test(slash): card mode on RadioInput
